Safari-611.1.10/LayoutTests/crypto/subtle/hmac-import-key-sign-sha224-expected.txt

Test signing with HMAC SHA-224 using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToHexString(signature) is expectedSignature
PASS successfullyParsed is true

TEST COMPLETE
